refactored require strings and variable names
consistency updates
