+++ services/commentServices/commentIndexService.js
@@ -1,4 +1,4 @@
-const commentModel = require('../../models/commentsModel.js')
+const commentModel = require('../../models/commentModel.js')
 
 //TODO So take the forums Id, and collect all related comments of that forum. Index it based on that forum id
 //based on the forum id
