+++ services/commentServices/commentUpdateService.js
@@ -1,4 +1,4 @@
-const commentModel = require('../../models/commentsModel.js')
+const commentModel = require('../../models/commentModel.js')
 
 const commentUpdateService = (commentId, commentBody) => {
         commentModel.findByIdAndUpdate(
