+++ services/tagsServices/tagShowService.js
@@ -1,4 +1,4 @@
-const tagModel = require('../../models/tagsModel.js')
+const tagModel = require('../../models/tagModel.js')
 //its odd that the error handling prevents a value from being returned.
 const tagShowService = (tagId) => {
     return tagModel.findById(tagId)
