+++ services/userServices/userCreateService.js
@@ -1,10 +1,10 @@
-const usersModel = require('../../models/usersModel.js')
+const usersModel = require('../../models/userModel.js')
 
 const userCreateService = async (userInput) => {
     try {
         return await usersModel.create(userInput)
     } catch (error) {
-        throw Error('Error while creating user. Location: userCreateService')
+        throw Error("Error while creating user. Location: userCreateService")
     }
 }
