+++ services/userServices/userLoginService.js
@@ -1,4 +1,4 @@
-const usersModel = require("../../models/usersModel");
+const usersModel = require('../../models/userModel.js')
 
 const userLoginService = async (userInput) => {
     try {
