+++ services/userServices/userShowService.js
@@ -1,10 +1,10 @@
-const usersModel = require('../../models/usersModel.js')
+const usersModel = require('../../models/userModel.js')
 
 const userShowService = async (userID) => {
     try {
         return await usersModel.findById(userID)
     } catch (error) {
-        throw Error('Error while fetching user. Location: userShowService')
+        throw Error("Error while fetching user. Location: userShowService")
     }
 }
