+++ services/userServices/userUpdateService.js
@@ -1,10 +1,10 @@
-const usersModel = require('../../models/usersModel')
+const usersModel = require('../../models/userModel.js')
 
 const userUpdateService = async (userIDAndInfo) => {
     try {
          return await usersModel.findByIdAndUpdate(userIDAndInfo)
     } catch (error) {
-        throw Error('Error while updating user. Location: userUpdateService')
+        throw Error("Error while updating user. Location: userUpdateService")
     }
 }
